requirements.txt
# Web
Django==5.1.1
django-cors-headers==4.4.0

# Data
pandas==2.2.2
openpyxl==3.1.5

# LLM integration (Ollama -> OpenAI -> template)
requests==2.32.3 
openai>=1.40.0
python-dotenv==1.0.1
